Changed for error handling: * Binding<T, U>: Meaning of T is now the typed of the original message for error handling, U is now the type for the produced/processed message out of T
Reason:
Let's assume T is String, and U is an Integer.
From usability a Binding<String, Integer> also shows the conversion from String->Integer. In case of any error, we will get the original String. In case of success, we will get the Integer. -> So this representation seems to be more intuitiv.